# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.Kermesse:
    properties:
      created_at:
        type: string
      description:
        type: string
      end_date:
        type: string
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      organisateurs:
        description: Relation many-to-many avec les organisateurs
        items:
          $ref: '#/definitions/models.User'
        type: array
      participants:
        description: Relation many-to-many avec les participants
        items:
          $ref: '#/definitions/models.User'
        type: array
      stands:
        description: Association one-to-many avec les stands
        items:
          $ref: '#/definitions/models.Stand'
        type: array
      start_date:
        type: string
      updated_at:
        type: string
    type: object
...
  models.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  models.PaymentRequest:
    properties:
      amount:
        description: Montant à payer en centimes
        type: integer
      currency:
        description: 'Devise (ex: ''EUR'')'
        type: string
      description:
        description: Description du paiement
        type: string
      source:
        description: 'Source de paiement (ex: un token de carte)'
        type: string
    type: object
  models.Stand:
    properties:
      created_at:
        description: Date de création du stand
        type: string
      description:
        description: Augmenté la taille pour plus de flexibilité
        type: string
      end_date:
        description: Date de fin du stand
        type: string
      id:
        type: integer
      kermesse:
        allOf:
        - $ref: '#/definitions/models.Kermesse'
        description: Associations
      kermesse_id:
        description: Référence vers la Kermesse
        type: integer
      location:
        description: Lieu où se trouve le stand
        type: string
      name:
        type: string
      owner:
        $ref: '#/definitions/models.User'
      owner_id:
        description: Référence vers le User qui tient le stand
        type: integer
      start_date:
        description: Date de début du stand
        type: string
      stock:
        type: integer
      type:
        description: e.g., "food", "game"
        type: string
      updated_at:
        description: Date de la dernière mise à jour du stand
        type: string
    type: object
...
  models.Transaction:
    properties:
      amount:
        description: Nombre de jetons utilisés
        type: integer
      currency:
        type: string
      id:
        type: integer
      kermesseID:
        type: integer
      paymentIntent:
        type: string
      stand:
        allOf:
        - $ref: '#/definitions/models.Stand'
        description: Associations
      standID:
        description: Référence vers le stand
        type: integer
      status:
        type: string
      timestamp:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        description: Référence vers l'utilisateur
        type: integer
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      passwordHash:
        type: string
      resetToken:
        description: Token de réinitialisation
        type: string
      resetTokenExpiry:
        description: Expiration du token de réinitialisation
        type: string
      role:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: contact@dratsisama.com
    name: Dr_AtsiSama
    url: http://dratsisama.com
  description: Documentation de l'API pour la gestion des kermesses scolaires avec
    fonctionnalités de stand, tombola, utilisateurs, etc.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  title: Kermisys API Documentation
  version: "1.0"
paths:
...
  /kermesses:
    get:
      description: Récupère la liste de toutes les kermesses avec pagination et filtres
      parameters:
      - description: Numéro de la page
        in: query
        name: page
        type: integer
      - description: Nombre d'éléments par page
        in: query
        name: limit
        type: integer
      - description: Filtrer par nom de la kermesse
        in: query
        name: name
        type: string
      - description: Filtrer par localisation de la kermesse
        in: query
        name: location
        type: string
      - description: 'Filtrer par date de début (format: YYYY-MM-DD)'
        in: query
        name: startDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Kermesse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Récupérer toutes les kermesses
      tags:
      - Kermesses
    post:
      consumes:
      - application/json
      description: Crée une nouvelle kermesse avec les détails fournis
      parameters:
      - description: Détails de la kermesse
        in: body
        name: kermesse
        required: true
        schema:
          $ref: '#/definitions/models.Kermesse'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Kermesse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Créer une nouvelle kermesse
      tags:
      - Kermesses
  /kermesses/{id}:
    delete:
      description: Supprime une kermesse par son ID
      parameters:
      - description: ID de la kermesse
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Supprimer une kermesse
      tags:
      - Kermesses
    get:
      description: Récupère les détails d'une kermesse par son ID
      parameters:
      - description: ID de la kermesse
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Kermesse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Récupérer une kermesse par ID
      tags:
      - Kermesses
    put:
      consumes:
      - application/json
      description: Met à jour les détails d'une kermesse existante
      parameters:
      - description: ID de la kermesse
        in: path
        name: id
        required: true
        type: integer
      - description: Détails mis à jour de la kermesse
        in: body
        name: kermesse
        required: true
        schema:
          $ref: '#/definitions/models.Kermesse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Kermesse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Mettre à jour une kermesse
      tags:
      - Kermesses
  /kermesses/{id}/participants/{user_id}:
    post:
      description: Ajoute un utilisateur en tant que participant à une kermesse
      parameters:
      - description: ID de la kermesse
        in: path
        name: id
        required: true
        type: integer
      - description: ID de l'utilisateur
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Ajouter un participant à une kermesse
      tags:
      - Kermesses
  /kermesses/{id}/stands/{stand_id}:
    post:
      description: Ajoute un stand à une kermesse
      parameters:
      - description: ID de la kermesse
        in: path
        name: id
        required: true
        type: integer
      - description: ID du stand
        in: path
        name: stand_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Ajouter un stand à une kermesse
      tags:
      - Kermesses
...
  /payments:
    post:
      consumes:
      - application/json
      description: Traite une demande de paiement en fonction des détails fournis
      parameters:
      - description: Données de paiement
        in: body
        name: paymentData
        required: true
        schema:
          $ref: '#/definitions/models.PaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Processus de paiement
      tags:
      - Paiements
  /payments/user/{userID}:
    get:
      description: Récupère toutes les transactions liées à un utilisateur
      parameters:
      - description: ID de l'utilisateur
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Transaction'
            type: array
      security:
      - Bearer: []
      summary: Récupérer les transactions d'un utilisateur
      tags:
      - Paiements
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"